Moved 2 `storage_t` methods not dependent on templates to `storage_base`
