+++ dev/get_prepared_statement.h
@@ -125,15 +125,15 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = internal::node_tuple_t<expression_type>;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
-        const result_tupe* result = nullptr;
+        using result_type = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
+        const result_type* result = nullptr;
         internal::iterate_ast(statement.expression, [&result, index = -1](auto& node) mutable {
             using node_type = std::decay_t<decltype(node)>;
             if(internal::is_bindable_v<node_type>) {
                 ++index;
             }
             if(index == N) {
-                internal::call_if_constexpr<std::is_same<result_tupe, node_type>::value>(
+                internal::call_if_constexpr<std::is_same<result_type, node_type>::value>(
                     [](auto& r, auto& n) {
                         r = &n;
                     },
@@ -150,8 +150,8 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = internal::node_tuple_t<expression_type>;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
-        result_tupe* result = nullptr;
+        using result_type = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
+        result_type* result = nullptr;
 
         internal::iterate_ast(statement.expression, [&result, index = -1](auto& node) mutable {
             using node_type = std::decay_t<decltype(node)>;
@@ -159,7 +159,7 @@ namespace sqlite_orm {
                 ++index;
             }
             if(index == N) {
-                internal::call_if_constexpr<std::is_same<result_tupe, node_type>::value>(
+                internal::call_if_constexpr<std::is_same<result_type, node_type>::value>(
                     [](auto& r, auto& n) {
                         r = const_cast<std::remove_reference_t<decltype(r)>>(&n);
                     },
