+++ include/sqlite_orm/sqlite_orm.h
@@ -8634,12 +8634,12 @@ namespace sqlite_orm {
 
     namespace internal {
         // Wrapper to reduce boiler-plate code
-        inline sqlite3_stmt* reset(sqlite3_stmt* stmt) {
+        inline sqlite3_stmt* reset_stmt(sqlite3_stmt* stmt) {
             sqlite3_reset(stmt);
             return stmt;
         }
 
-        // note: query is taken by value, such that it is thrown away early
+        // note: query is deliberately taken by value, such that it is thrown away early
         inline sqlite3_stmt* prepare_stmt(sqlite3* db, std::string query) {
             sqlite3_stmt* stmt;
             if(sqlite3_prepare_v2(db, query.c_str(), -1, &stmt, nullptr) != SQLITE_OK) {
@@ -13663,8 +13663,6 @@ namespace sqlite_orm {
     }
 }
 
-// #include "statement_finalizer.h"
-
 // #include "row_extractor.h"
 
 // #include "connection_holder.h"
@@ -13991,15 +13989,25 @@ namespace sqlite_orm {
           protected:
             void rename_table(sqlite3* db, const std::string& oldName, const std::string& newName) const {
                 std::stringstream ss;
-                ss << "ALTER TABLE " << quote_identifier(oldName) << " RENAME TO " << quote_identifier(newName)
+                ss << "ALTER TABLE " << streaming_identifier(oldName) << " RENAME TO " << streaming_identifier(newName)
                    << std::flush;
                 perform_void_exec(db, ss.str());
+            }
+
+            /**
+             *  Checks whether table exists in db. Doesn't check storage itself - works only with actual database.
+             *  Note: table can be not mapped to a storage
+             *  @return true if table with a given name exists in db, false otherwise.
+             */
+            bool table_exists(const std::string& tableName) {
+                auto con = this->get_connection();
+                return this->table_exists(con.get(), tableName);
             }
 
             bool table_exists(sqlite3* db, const std::string& tableName) const {
                 bool result = false;
                 std::stringstream ss;
-                ss << "SELECT COUNT(*) FROM sqlite_master WHERE type = " << quote_string_literal("table")
+                ss << "SELECT COUNT(*) FROM sqlite_master WHERE type = " << streaming_identifier("table")
                    << " AND name = " << quote_string_literal(tableName) << std::flush;
                 perform_exec(
                     db,
@@ -14013,6 +14021,16 @@ namespace sqlite_orm {
                     },
                     &result);
                 return result;
+            }
+
+            void add_generated_cols(std::vector<const table_xinfo*>& columnsToAdd,
+                                    const std::vector<table_xinfo>& storageTableInfo) {
+                //  iterate through storage columns
+                for(const table_xinfo& storageColumnInfo: storageTableInfo) {
+                    if(storageColumnInfo.hidden) {
+                        columnsToAdd.push_back(&storageColumnInfo);
+                    }
+                }
             }
 
           public:
@@ -14043,9 +14061,9 @@ namespace sqlite_orm {
             }
 
             /**
-             *  Returns libsqltie3 lib version, not sqlite_orm
+             *  Returns libsqlite3 version, not sqlite_orm
              */
-            std::string libversion() {
+            static std::string libversion() {
                 return sqlite3_libversion();
             }
 
@@ -17296,15 +17314,7 @@ namespace sqlite_orm {
                 perform_void_exec(db, ss.str());
             }
 #endif
-            void add_generated_cols(std::vector<const table_xinfo*>& columnsToAdd,
-                                    const std::vector<table_xinfo>& storageTableInfo) {
-                //  iterate through storage columns
-                for(const table_xinfo& storageColumnInfo: storageTableInfo) {
-                    if(storageColumnInfo.hidden) {
-                        columnsToAdd.push_back(&storageColumnInfo);
-                    }
-                }
-            }
+
             template<class I>
             void drop_create_with_loss(sqlite3* db, const I& tImpl) {
                 // eliminated all transaction handling
@@ -17559,7 +17569,7 @@ namespace sqlite_orm {
              */
             template<class O, class... Ids>
             std::shared_ptr<O> get_no_throw(Ids... ids) {
-                return std::shared_ptr<O>(get_pointer<O>(std::forward<Ids>(ids)...));
+                return std::shared_ptr<O>(this->get_pointer<O>(std::forward<Ids>(ids)...));
             }
 
 #ifdef SQLITE_ORM_OPTIONAL_SUPPORTED
@@ -18105,7 +18115,7 @@ namespace sqlite_orm {
                                           bool preserve);
 
             template<class C>
-            void add_column(const std::string& tableName, const C& column, sqlite3* db) const {
+            void add_column(sqlite3* db, const std::string& tableName, const C& column) const {
                 using context_t = serializer_context<impl_type>;
 
                 context_t context{this->impl};
@@ -18178,16 +18188,6 @@ namespace sqlite_orm {
                     result.insert({tImpl.table.name, schemaStatus});
                 });
                 return result;
-            }
-
-            /**
-             *  Checks whether table exists in db. Doesn't check storage itself - works only with actual database.
-             *  Note: table can be not mapped to a storage
-             *  @return true if table with a given name exists in db, false otherwise.
-             */
-            bool table_exists(const std::string& tableName) {
-                auto con = this->get_connection();
-                return this->table_exists(con.get(), tableName);
             }
 
             using storage_base::table_exists;  // now that it is in storage_base make it into overload set
@@ -18304,14 +18304,14 @@ namespace sqlite_orm {
 
             template<class... Args>
             void execute(const prepared_statement_t<replace_raw_t<Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 iterate_ast(statement.expression.args, conditional_binder{statement.stmt});
                 perform_step(stmt);
             }
 
             template<class... Args>
             void execute(const prepared_statement_t<insert_raw_t<Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 iterate_ast(statement.expression.args, conditional_binder{stmt});
                 perform_step(stmt);
             }
@@ -18322,7 +18322,7 @@ namespace sqlite_orm {
                 using expression_type = typename statement_type::expression_type;
                 using object_type = typename expression_object_type<expression_type>::type;
 
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 tuple_value_binder{stmt}(
                     statement.expression.columns.columns,
@@ -18340,7 +18340,7 @@ namespace sqlite_orm {
                 using expression_type = typename statement_type::expression_type;
                 using object_type = typename expression_object_type<expression_type>::type;
 
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 auto processObject = [&tImpl = this->get_impl<object_type>(),
                                       bind_value = field_value_binder{stmt}](auto& object) mutable {
@@ -18374,7 +18374,7 @@ namespace sqlite_orm {
                 using expression_type = typename statement_type::expression_type;
                 using object_type = typename expression_object_type<expression_type>::type;
 
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 auto processObject = [&tImpl = this->get_impl<object_type>(),
                                       bind_value = field_value_binder{stmt}](auto& object) mutable {
@@ -18411,7 +18411,7 @@ namespace sqlite_orm {
 
             template<class T, class... Ids>
             void execute(const prepared_statement_t<remove_t<T, Ids...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 iterate_ast(statement.expression.ids, conditional_binder{stmt});
                 perform_step(stmt);
             }
@@ -18422,7 +18422,7 @@ namespace sqlite_orm {
                 using expression_type = typename statement_type::expression_type;
                 using object_type = typename expression_object_type<expression_type>::type;
 
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 auto& tImpl = this->get_impl<object_type>();
 
                 field_value_binder bind_value{stmt};
@@ -18444,7 +18444,7 @@ namespace sqlite_orm {
 
             template<class T, class... Ids>
             std::unique_ptr<T> execute(const prepared_statement_t<get_pointer_t<T, Ids...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression.ids, conditional_binder{stmt});
 
@@ -18460,7 +18460,7 @@ namespace sqlite_orm {
 #ifdef SQLITE_ORM_OPTIONAL_SUPPORTED
             template<class T, class... Ids>
             std::optional<T> execute(const prepared_statement_t<get_optional_t<T, Ids...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression.ids, conditional_binder{stmt});
 
@@ -18475,7 +18475,7 @@ namespace sqlite_orm {
 
             template<class T, class... Ids>
             T execute(const prepared_statement_t<get_t<T, Ids...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression.ids, conditional_binder{stmt});
 
@@ -18511,14 +18511,14 @@ namespace sqlite_orm {
 
             template<class T, class... Args>
             void execute(const prepared_statement_t<remove_all_t<T, Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 iterate_ast(statement.expression.conditions, conditional_binder{stmt});
                 perform_step(stmt);
             }
 
             template<class... Args, class... Wargs>
             void execute(const prepared_statement_t<update_all_t<set_t<Args...>, Wargs...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
                 conditional_binder bind_node{stmt};
                 iterate_tuple(statement.expression.set.assigns, [&bind_node](auto& setArg) {
                     iterate_ast(setArg, bind_node);
@@ -18529,7 +18529,7 @@ namespace sqlite_orm {
 
             template<class T, class... Args, class R = column_result_of_t<self, T>>
             std::vector<R> execute(const prepared_statement_t<select_t<T, Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression, conditional_binder{stmt});
 
@@ -18544,7 +18544,7 @@ namespace sqlite_orm {
 
             template<class T, class R, class... Args>
             R execute(const prepared_statement_t<get_all_t<T, R, Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression, conditional_binder{stmt});
 
@@ -18560,7 +18560,7 @@ namespace sqlite_orm {
 
             template<class T, class R, class... Args>
             R execute(const prepared_statement_t<get_all_pointer_t<T, R, Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression, conditional_binder{stmt});
 
@@ -18577,7 +18577,7 @@ namespace sqlite_orm {
 #ifdef SQLITE_ORM_OPTIONAL_SUPPORTED
             template<class T, class R, class... Args>
             R execute(const prepared_statement_t<get_all_optional_t<T, R, Args...>>& statement) {
-                sqlite3_stmt* stmt = reset(statement.stmt);
+                sqlite3_stmt* stmt = reset_stmt(statement.stmt);
 
                 iterate_ast(statement.expression, conditional_binder{stmt});
 
@@ -19120,15 +19120,15 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = internal::node_tuple_t<expression_type>;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
-        const result_tupe* result = nullptr;
+        using result_type = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
+        const result_type* result = nullptr;
         internal::iterate_ast(statement.expression, [&result, index = -1](auto& node) mutable {
             using node_type = std::decay_t<decltype(node)>;
             if(internal::is_bindable_v<node_type>) {
                 ++index;
             }
             if(index == N) {
-                internal::call_if_constexpr<std::is_same<result_tupe, node_type>::value>(
+                internal::call_if_constexpr<std::is_same<result_type, node_type>::value>(
                     [](auto& r, auto& n) {
                         r = &n;
                     },
@@ -19145,8 +19145,8 @@ namespace sqlite_orm {
         using expression_type = typename statement_type::expression_type;
         using node_tuple = internal::node_tuple_t<expression_type>;
         using bind_tuple = typename internal::bindable_filter<node_tuple>::type;
-        using result_tupe = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
-        result_tupe* result = nullptr;
+        using result_type = std::tuple_element_t<static_cast<size_t>(N), bind_tuple>;
+        result_type* result = nullptr;
 
         internal::iterate_ast(statement.expression, [&result, index = -1](auto& node) mutable {
             using node_type = std::decay_t<decltype(node)>;
@@ -19154,7 +19154,7 @@ namespace sqlite_orm {
                 ++index;
             }
             if(index == N) {
-                internal::call_if_constexpr<std::is_same<result_tupe, node_type>::value>(
+                internal::call_if_constexpr<std::is_same<result_type, node_type>::value>(
                     [](auto& r, auto& n) {
                         r = const_cast<std::remove_reference_t<decltype(r)>>(&n);
                     },
@@ -19460,7 +19460,7 @@ namespace sqlite_orm {
                                     if(column.name != colInfo->name) {
                                         return;
                                     }
-                                    this->add_column(tImpl.table.name, column, db);
+                                    this->add_column(db, tImpl.table.name, column);
                                 });
                             }
                             res = sync_schema_result::new_columns_added;
